pool/microservicio/dominio/src/main/java/com/ceiba/cliente/servicio/ServicioCrearCliente.java
package com.ceiba.cliente.servicio;

import com.ceiba.cliente.modelo.entidad.Cliente;
import com.ceiba.cliente.puerto.repositorio.RepositorioCliente;

public class ServicioCrearCliente {

    private final RepositorioCliente repositorioCliente;

    public ServicioCrearCliente(RepositorioCliente repositorioCliente) {
        this.repositorioCliente = repositorioCliente;
    }

    public void ejecutar(Cliente cliente) {
        var existe = this.repositorioCliente.existePorIdentificacion(cliente.getIdentificacion());

        if(!existe) {
            this.repositorioCliente.crear(cliente);
        }
    }

}
